requirements.txt
Django==5.0.7
djangorestframework==3.14.0
drf-spectacular==0.27.0